feat: move calculator layout to its own file
This is update to ensure the routing can be changed anytime (like from framework mode to declarative, etc)
